wip splitted out bufr export test
